+++ CodonU/vizualizer/plot_mca_codon_freq.py
@@ -4,10 +4,12 @@ from prince import PCA
 from Bio.SeqIO import parse
 from CodonU.analyzer.internal_comp import filter_reference
 from CodonU.correspondence_analysis.mca_codon_freq import mca_codon_freq
+from CodonU.file_handler import make_dir
+from CodonU.file_handler.internal_comp import is_file_writeable
 
 
 def plot_mca_codon_freq(handle: str, genetic_table_num: int, min_len_threshold: int = 200, n_components: int = 59,
-                        organism_name: str | None = None, save_image: bool = False, folder_path: str = ''):
+                        organism_name: str | None = None, save_image: bool = False, folder_path: str = 'Report'):
     """
     Plots the principal component analysis based on codon frequency
 
@@ -43,8 +45,10 @@ def plot_mca_codon_freq(handle: str, genetic_table_num: int, min_len_threshold: 
     sup_title = f'Multivariate analysis of Codon Frequency of {organism_name}' if organism_name else 'Multivariate analysis of Codon Frequency'
     plt.suptitle(sup_title)
     if save_image:
+        make_dir(folder_path)
         name = f'Multivariate_analysis_codon_freq_{organism_name}.png' if organism_name else 'Multivariate_analysis_codon_freq.png'
         file_name = join(folder_path, name)
-        plt.savefig(file_name, dpi=500)
+        if is_file_writeable(file_name):
+            plt.savefig(file_name, dpi=500)
     plt.show()
     plt.close()
